rename: Swagger 적용을 위한 인터페이스명을 Spec으로 수정
 - 각 도메인 패키지 하위의 api 패키지 삭제
 - Swagger 적용을 위한 인터페이스명을 Spec으로 바꾸고 도메인 하위로 이동
