+++ src/main/java/kr/hhplus/be/server/interfaces/api/product/ProductMockController.java
@@ -1,16 +1,15 @@
 package kr.hhplus.be.server.interfaces.api.product;
 
-import kr.hhplus.be.server.interfaces.api.product.api.ProductApi;
+import kr.hhplus.be.server.common.response.ApiResponse;
 import kr.hhplus.be.server.interfaces.api.product.dto.response.ProductListResponse;
 import kr.hhplus.be.server.interfaces.api.product.dto.response.ProductResponse;
-import kr.hhplus.be.server.common.response.ApiResponse;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 
 @RestController
-public class ProductMockController implements ProductApi {
+public class ProductMockController implements ProductSpec {
 
     @GetMapping("/api/v1/products")
     public ApiResponse<ProductListResponse<ProductResponse>> getProducts() {
